feat: add trigger threshold and enable min threshold
