+++ internal/analysis/processor/processor.go
@@ -280,21 +280,6 @@ func (p *Processor) processResults(item *queue.Results) []Detections {
 	return detections
 }
 
-// addSpeciesToDynamicThresholds adds a species to the dynamic thresholds map if it doesn't already exist.
-func (p *Processor) addSpeciesToDynamicThresholds(speciesLowercase string, baseThreshold float32) {
-	_, exists := p.DynamicThresholds[speciesLowercase]
-	if !exists {
-		log.Printf("Initializing dynamic threshold for %s\n", speciesLowercase)
-		p.DynamicThresholds[speciesLowercase] = &DynamicThreshold{
-			Level:         0,
-			CurrentValue:  float64(baseThreshold),
-			Timer:         time.Now(),
-			HighConfCount: 0,
-			ValidHours:    1, // Default to 1 hour, can be configured
-		}
-	}
-}
-
 // handleDogDetection handles the detection of dog barks and updates the last detection timestamp.
 func (p *Processor) handleDogDetection(item *queue.Results, speciesLowercase string, result datastore.Results) {
 	if p.Settings.Realtime.DogBarkFilter.Enabled && strings.Contains(speciesLowercase, "dog") && result.Confidence > p.Settings.Realtime.DogBarkFilter.Confidence {
@@ -311,17 +296,6 @@ func (p *Processor) handleHumanDetection(item *queue.Results, speciesLowercase s
 	}
 }
 
-// getConfidenceThreshold retrieves the confidence threshold for a species, using custom or global thresholds.
-func (p *Processor) getConfidenceThreshold(speciesLowercase string) float32 {
-	confidenceThreshold, exists := p.SpeciesConfig.Threshold[speciesLowercase]
-	if !exists {
-		confidenceThreshold = float32(p.Settings.BirdNET.Threshold)
-	} else if p.Settings.Debug {
-		log.Printf("\nUsing confidence threshold of %.2f for %s\n", confidenceThreshold, speciesLowercase)
-	}
-	return confidenceThreshold
-}
-
 // getBaseConfidenceThreshold retrieves the confidence threshold for a species, using custom or global thresholds.
 func (p *Processor) getBaseConfidenceThreshold(speciesLowercase string) float32 {
 	confidenceThreshold, exists := p.SpeciesConfig.Threshold[speciesLowercase]
@@ -331,57 +305,6 @@ func (p *Processor) getBaseConfidenceThreshold(speciesLowercase string) float32 
 		log.Printf("\nUsing confidence threshold of %.2f for %s\n", confidenceThreshold, speciesLowercase)
 	}
 	return confidenceThreshold
-}
-
-// getAdjustedConfidenceThreshold applies dynamic threshold logic to adjust the confidence threshold based on recent detections.
-func (p *Processor) getAdjustedConfidenceThreshold(speciesLowercase string, result datastore.Results, baseThreshold float32) float32 {
-	dt, exists := p.DynamicThresholds[speciesLowercase]
-	if !exists {
-		return baseThreshold
-	}
-
-	// If the detection confidence exceeds the base threshold
-	if result.Confidence > baseThreshold {
-		log.Printf("High confidence detection of %s\n", speciesLowercase)
-		dt.HighConfCount++
-		dt.Timer = time.Now().Add(time.Duration(dt.ValidHours) * time.Hour)
-
-		// Adjust the dynamic threshold based on the number of high-confidence detections
-		switch dt.HighConfCount {
-		case 1:
-			dt.Level = 1
-			dt.CurrentValue = float64(baseThreshold * 0.75)
-		case 2:
-			dt.Level = 2
-			dt.CurrentValue = float64(baseThreshold * 0.5)
-		case 3:
-			dt.Level = 3
-			dt.CurrentValue = float64(baseThreshold * 0.25)
-		}
-
-		log.Printf("Applying dynamic threshold for %s to %.2f", speciesLowercase, dt.CurrentValue)
-	} else if time.Now().After(dt.Timer) {
-		//log.Printf("Dynamic threshold timer expired for %s\n", speciesLowercase)
-		// Reset the dynamic threshold if the timer has expired
-		dt.Level = 0
-		dt.CurrentValue = float64(baseThreshold)
-		dt.HighConfCount = 0
-		//log.Printf("Reset dynamic threshold for %s", speciesLowercase)
-	}
-	return float32(dt.CurrentValue)
-}
-
-// cleanUpDynamicThresholds removes stale dynamic thresholds for species that haven't been detected for a long time.
-func (p *Processor) cleanUpDynamicThresholds() {
-	staleDuration := 24 * time.Hour // Duration after which a dynamic threshold is considered stale
-	now := time.Now()
-
-	for species, dt := range p.DynamicThresholds {
-		if now.Sub(dt.Timer) > staleDuration {
-			log.Printf("Removing stale dynamic threshold for %s", species)
-			delete(p.DynamicThresholds, species)
-		}
-	}
 }
 
 func (p *Processor) generateClipName(scientificName string, confidence float32) string {
