+++ internal/conf/defaults.go
@@ -48,6 +48,7 @@ func setDefaultConfig() {
 
 	viper.SetDefault("realtime.dynamicthreshold.enabled", true)
 	viper.SetDefault("realtime.dynamicthreshold.debug", false)
+	viper.SetDefault("realtime.dynamicthreshold.trigger", 0.90)
 	viper.SetDefault("realtime.dynamicthreshold.min", 0.20)
 	viper.SetDefault("realtime.dynamicthreshold.validhours", 24)
